Move sync-related functionality to module
